Moved ModelSet tests into separate file
