+++ tests/model/test_sos_model.py
@@ -14,7 +14,7 @@ from smif.metadata import MetadataSet
 from smif.model.dependency import Dependency
 from smif.model.scenario_model import ScenarioModel
 from smif.model.sector_model import SectorModel
-from smif.model.sos_model import ModelSet, SosModel, SosModelBuilder
+from smif.model.sos_model import SosModel, SosModelBuilder
 
 from ..fixtures.water_supply import WaterSupplySectorModel
 
@@ -315,73 +315,6 @@ class TestSosModel():
                 }
 
         sos_model.simulate(2010, data)
-
-
-class TestIterations:
-
-    def test_guess_outputs_zero(self, get_sector_model_object):
-        """If no previous timestep has results, guess outputs as zero
-        """
-        ws_model = get_sector_model_object
-        model_set = ModelSet([ws_model])
-
-        results = model_set.guess_results(ws_model, 2010, {})
-        expected = {
-            "cost": np.zeros((1, 1)),
-            "water": np.zeros((1, 1))
-        }
-        assert results == expected
-
-    def test_guess_outputs_last_year(self, get_sector_model_object):
-        """If a previous timestep has results, guess outputs as identical
-        """
-        ws_model = get_sector_model_object
-        model_set = ModelSet([ws_model])
-
-        expected = {
-            "cost": np.array([[3.14]]),
-            "water": np.array([[2.71]])
-        }
-
-        # set up data as though from previous timestep simulation
-        data = {
-            2010: {
-                'water_supply': expected
-            },
-            2011: {}
-        }
-
-        results = model_set.guess_results(ws_model, 2011, data)
-        assert results == expected
-
-    def test_converged_first_iteration(self, get_sector_model_object):
-        """Should not report convergence after a single iteration
-        """
-        ws_model = get_sector_model_object
-        model_set = ModelSet([ws_model])
-
-        results = model_set.guess_results(ws_model, 2010, {})
-        model_set.iterated_results = [{ws_model.name: results}]
-
-        assert not model_set.converged()
-
-    def test_converged_two_identical(self, get_sector_model_object):
-        """Should report converged if the last two output sets are identical
-        """
-        ws_model = get_sector_model_object
-        model_set = ModelSet([ws_model])
-
-        results = model_set.guess_results(ws_model, 2010, {})
-        model_set.iterated_results = [
-            {
-                "water_supply": results
-            },
-            {
-                "water_supply": results
-            }
-        ]
-
-        assert model_set.converged()
 
 
 @fixture(scope='function')
